+++ lib/views/profile/profile_screen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:iconsax/iconsax.dart';
 
@@ -7,6 +8,9 @@ import 'package:iconsax/iconsax.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_facebook_auth/flutter_facebook_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
+
+// Utils
+import 'package:recycle/utils/providers/user_data_provider.dart';
 
 // Views
 import 'package:recycle/views/profile/edit_profile_screen.dart';
@@ -17,19 +21,23 @@ import 'widgets/profile_widget.dart';
 // Constants
 import 'package:recycle/constants.dart';
 
-class ProfilePage extends StatefulWidget {
+class ProfilePage extends ConsumerStatefulWidget {
   const ProfilePage({Key? key, required this.title}) : super(key: key);
 
   final String title;
 
   @override
-  State<StatefulWidget> createState() => ProfilePageState();
+  ConsumerState<ConsumerStatefulWidget> createState() => ProfilePageState();
 }
 
-class ProfilePageState extends State<ProfilePage> {
+class ProfilePageState extends ConsumerState<ProfilePage> {
   bool dark = false;
   @override
   Widget build(BuildContext context) {
+    // Automatically refresh the data whenever this widget is built
+    ref.read(userDataNotifier.notifier).refreshData();
+    final profileData = ref.watch(userDataNotifier);
+
     double windowWidth = MediaQuery.of(context).size.width;
     double windowHeight = MediaQuery.of(context).size.height;
 
@@ -73,12 +81,13 @@ class ProfilePageState extends State<ProfilePage> {
                 ),
                 child: Center(
                   child: ListTile(
-                    leading: const CircleAvatar(
+                    leading: CircleAvatar(
                       radius: 30,
-                      backgroundImage: NetworkImage(
-                          'https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_1280.png'),
+                      backgroundImage: NetworkImage(profileData
+                              .profilePictureUrl ??
+                          "https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_1280.png"),
                     ),
-                    title: const Text("John Doe"),
+                    title: Text(profileData.profileUsername ?? "J. Doe"),
                     subtitle: const Text("johndoe@gmail.com"),
                     trailing: IconButton(
                       color: primaryGreen,
@@ -125,7 +134,8 @@ class ProfilePageState extends State<ProfilePage> {
                               color: primaryGreen,
                             ),
                             SizedBox(width: windowWidth * 0.02),
-                            const Text("1200"),
+                            // NOTE: might be wrong
+                            Text(profileData.points.toString()),
                           ],
                         ),
                         SizedBox(height: windowHeight * 0.01),
